+++ src/britive/policies.py
@@ -1,12 +1,10 @@
-from . import exceptions
-import datetime
-import json
+from .system.policies import SystemPolicies
 
 
 class Policies:
     def __init__(self, britive):
         self.britive = britive
-        self.base_url = f'{self.britive.base_url}/v1/policy-admin'
+        self.base_url = f'{self.britive.base_url}/v1/policy-admin/policies'
 
     @staticmethod
     def build(name: str, description: str = '', draft: bool = False, active: bool = True,
@@ -14,22 +12,22 @@ class Policies:
               service_identities: list = None, permissions: list = None, roles: list = None, ips: list = None,
               from_time: str = None, to_time: str = None, approval_notification_medium: str = None,
               time_to_approve: int = 5, access_validity_time: int = 120, approver_users: list = None,
-              approver_tags: list = None, access_type: str = 'Allow') -> dict:
+              approver_tags: list = None, access_type: str = 'Allow', identifier_type: str = 'name') -> dict:
         """
         Build a policy document given the provided inputs.
 
         :param name: The name of the policy.
         :param description: An optional description of the policy.
         :param draft: Indicates if the policy is a draft. Defaults to `False`.
-        :param active: Indicates if the policy is a active. Defaults to `True`.
+        :param active: Indicates if the policy is active. Defaults to `True`.
         :param read_only: Indicates if the policy is a read only. Defaults to `False`.
-        :param users: Optional list of user names to which this policy applies.
-        :param tags: Optional list of tag names to which this policy applies.
-        :param tokens: Optional list of token IDs to which this policy applies.
-        :param service_identities: Optional list of service identity names to which this policy applies.
-        :param permissions: Optional list of permission names this policy grants. Provide either this parameter
+        :param users: Optional list of user names or ids to which this policy applies.
+        :param tags: Optional list of tag names or ids to which this policy applies.
+        :param tokens: Optional list of token names or ids to which this policy applies.
+        :param service_identities: Optional list of service identity names or ids to which this policy applies.
+        :param permissions: Optional list of permission names or ids this policy grants. Provide either this parameter
             or `roles`.
-        :param roles: Optional list of role names to which this policy applies. Provider either this parameter
+        :param roles: Optional list of role names or ids to which this policy applies. Provider either this parameter
             or `permissions`.
         :param ips: Optional list of IP addresses for which this policy applies. Provide in CIDR notation
             or dotted decimal format for individual (/32) IP addresses.
@@ -47,83 +45,40 @@ class Policies:
             to 5 minutes.
         :param access_validity_time: Optional number of minutes the access is valid after approval. Defaults to 120
             minutes.
-        :param approver_users: Optional list of user names who are to be considered approvers.
+        :param approver_users: Optional list of user names or ids who are to be considered approvers.
             If `approval_notification_medium` is set then either `approver_users` or `approver_tags` is required.
         :param approver_tags: Optional list of tag names who are considered approvers.
             If `approval_notification_medium` is set then either `approver_users` or `approver_tags` is required.
         :param access_type: The type of access this policy provides. Valid values are `Allow` and `Deny`. Defaults
             to `Allow`.
-        :return: A dict which can be provided as a profile policy to `create` and `update`.
+        :param identifier_type: Valid values are `id` or `name`. Defaults to `name`. Represents which type of
+            identifiers are being provided to the other parameters. Either all identifiers must be names or all
+            identifiers must be IDs.
+        :return: A dict which can be provided as a policy to `create` and `update`.
         """
 
-        condition = {}
+        return SystemPolicies.build(
+            name=name,
+            description=description,
+            draft=draft,
+            active=active,
+            read_only=read_only,
+            users=users,
+            tags=tags,
+            tokens=tokens,
+            service_identities=service_identities,
+            permissions=permissions,
+            roles=roles,
+            ips=ips,
+            from_time=from_time,
+            to_time=to_time,
+            approval_notification_medium=approval_notification_medium,
+            time_to_approve=time_to_approve,
+            access_validity_time=access_validity_time,
+            approver_users=approver_users,
+            approver_tags=approver_tags,
+            access_type=access_type,
+            identifier_type=identifier_type
+        )
 
-        # handle ip address logic
-        if ips:
-            condition['ipAddress'] = ','.join(ips)
 
-        # handle from_time and to_time logic
-        if from_time and not to_time:
-            raise ValueError('if from_time is provided then to_time must also be provided.')
-        if to_time and not from_time:
-            raise ValueError('if to_time is provided then from_time must also be provided.')
-        if from_time and to_time:
-            condition['timeOfAccess'] = {
-                'from': from_time,
-                'to': to_time
-            }
-
-        # handle approval logic
-        if approval_notification_medium:
-            if not approver_users and not approver_tags:
-                raise ValueError('when approval is required either approver_tags or approver_users or both '
-                                 'must be provided')
-            approval_condition = {
-                'notificationMedium': approval_notification_medium,
-                'timeToApprove': time_to_approve,
-                'validFor': access_validity_time,
-                'approvers': {
-                    'userIds': approver_users,
-                    'tags': approver_tags
-                }
-            }
-
-            if not approver_users:
-                approval_condition['approvers'].pop('userIds')
-            if not approver_tags:
-                approval_condition['approvers'].pop('tags')
-
-            condition['approval'] = approval_condition
-
-        # put it all together
-        policy = {
-            'name': name,
-            'description': description,
-            'isActive': active,
-            'isDraft': draft,
-            'isReadOnly': read_only,
-            'accessType': access_type,
-            'members': {
-                'users': [{'name': u} for u in users] if users else None,
-                'tags': [{'name': t} for t in tags] if tags else None,
-                'serviceIdentities': [{'name': s} for s in service_identities] if service_identities else None,
-                'tokens': [{'name': t} for t in tokens] if tokens else None
-            },
-            'condition': json.dumps(condition, default=str)
-        }
-
-        if not users:
-            policy['members'].pop('users')
-        if not tags:
-            policy['members'].pop('tags')
-        if not service_identities:
-            policy['members'].pop('serviceIdentities')
-        if not tokens:
-            policy['members'].pop('tokens')
-
-        if permissions:
-            policy['permissions'] = [{'name': p} for p in permissions]
-        if roles:
-            policy['roles'] = [{'name': r} for r in roles]
-
-        return policy