+++ tests/test_100-permissions.py
@@ -1,6 +1,7 @@
 from .cache import *  # will also import some globals like `britive`
 
 
+@pytest.mark.skipif(scan_skip, reason=scan_skip_message)
 def test_list(cached_application, cached_environment):
     permissions = britive.permissions.list(
         application_id=cached_application['appContainerId'],
@@ -11,6 +12,7 @@ def test_list(cached_application, cached_environment):
     assert isinstance(permissions[0], dict)
 
 
+@pytest.mark.skipif(scan_skip, reason=scan_skip_message)
 def test_accounts(cached_application, cached_environment, cached_permission):
     accounts = britive.permissions.accounts(
         permission_id=cached_permission['appPermissionId'],
@@ -21,6 +23,7 @@ def test_accounts(cached_application, cached_environment, cached_permission):
     assert isinstance(accounts, list)
 
 
+@pytest.mark.skipif(scan_skip, reason=scan_skip_message)
 def test_groups(cached_application, cached_environment, cached_permission):
     groups = britive.permissions.groups(
         permission_id=cached_permission['appPermissionId'],